third_party/blink/web_tests/platform/generic/fast/html/draggable-expected.txt

Intrinsic
 webkit.org <a no href> <span>
draggable attr set to true
 webkit.org <a no href> <span>
draggable attr set to false
 webkit.org <a no href> <span>
draggable attr set to maybe
 webkit.org <a no href> <span>
draggable dom prop set to true
 webkit.org <a no href> <span>
draggable dom prop set to false
 webkit.org <a no href> <span>
dom - attr
----------
true - null
true - null
false - null
false - null
true - true
true - true
true - true
true - true
false - false
false - false
false - false
false - false
true - maybe
true - maybe
false - maybe
false - maybe
true - true
true - true
true - true
true - true
false - false
false - false
false - false
false - false
